requirements.txt
aiohttp==2.2.5
aiohttp-cors==0.5.3
aiomysql==0.0.11
gunicorn==19.4.5
PyJWT==1.4.0
